src/auth.py
import os.path

from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow

SCOPES = ['https://mail.google.com/']


def auth():
    """Shows basic usage of the People API.
    Prints the name of the first 10 connections.
    """
    creds = None
    # The file token.json stores the user's access and refresh tokens, and is
    # created automatically when the authorization flow completes for the first
    # time.
    if os.path.exists('token.json'):
        creds = Credentials.from_authorized_user_file('token.json', SCOPES)
    # If there are no (valid) credentials available, let the user log in.

    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                'credentials.json', SCOPES)
            creds = flow.run_local_server(port=0)

        # Save the credentials for the next run
        # TODO: still figuring how to handle this as lambda is read only so can't write to file and this logic will happen on frontend as well
        # TEMP CHANGE
        # Don't try and write creds, just reutnr them as lambda doesn't allow file write permission
        #####
        # with open('token.json', 'w') as token:
        #     token.write(creds.to_json())
        #####

    # WATCH THIS FUCKING INDENT
    return creds
